Remove old Table and rename List to Table
