+++ src/components/screens/delegates/delegatesTable/delegatesTable.js
@@ -10,7 +10,7 @@ import BoxHeader from '../../../toolbox/box/header';
 import BoxContent from '../../../toolbox/box/content';
 import BoxTabs from '../../../toolbox/tabs';
 import { Input } from '../../../toolbox/inputs';
-import Table from '../../../toolbox/list';
+import Table from '../../../toolbox/table';
 import DelegateRow from './delegateRow';
 import header from './tableHeader';
