+++ src/components/toolbox/table/loading.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { isReactComponent } from '../../../utils/helpers';
-import styles from '../table/table.css';
+import styles from './table.css';
 
 const Loading = ({ Element, headerInfo }) => (
   isReactComponent(Element)
